try to fix lint
